addition of sub schema support

+++ pkg/schemas/store.go
@@ -15,7 +15,9 @@ type ReadableStore interface {
 	storage.ReadableStorage
 }
 
+// Store implementation to abstract store operations
 type readStoreImpl struct {
+	cache map[string][]byte
 	shell *shell.Shell
 }
 
@@ -36,10 +38,17 @@ func (rs readStoreImpl) Has(ctx context.Context, key string) (bool, error) {
 }
 
 func (rs readStoreImpl) Get(ctx context.Context, key string) ([]byte, error) {
+	if rs.cache == nil {
+		rs.cache = make(map[string][]byte)
+	}
+
+	if rs.cache[key] != nil {
+		return rs.cache[key], nil
+	}
+
 	time_stamp := fmt.Sprintf("%d", time.Now().Unix())
 
 	out_path := filepath.Join(os.TempDir(), key+time_stamp+".txt")
-
 	defer os.Remove(out_path)
 
 	err := rs.shell.Get(key, out_path)
@@ -53,6 +62,8 @@ func (rs readStoreImpl) Get(ctx context.Context, key string) ([]byte, error) {
 	if err != nil {
 		return nil, err
 	}
+
+	rs.cache[key] = buf
 
 	return buf, nil
 }
